reclass nodes by splitting image and image collection
